% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coords.R
\name{as.coords}
\alias{as.coords}
\title{Coord class}
\usage{
as.coords(place, box, point)
}
\arguments{
\item{place}{Name of the location}

\item{box}{Bounding box of the location}

\item{point}{lon/lat of the address}
}
\description{
Return an object containing place, box, point to be used with
other rtweet functions.
}
