% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_twitter.R
\name{connect_to_api}
\alias{connect_to_api}
\title{Connect to Twitter API}
\usage{
connect_to_api(creds_list)
}
\arguments{
\item{creds_list}{A list of Twitter API credentials}
}
\description{
Connect to Twitter API using the credentials provided.
}
