% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tidytext.R
\name{filter_filler_words}
\alias{filter_filler_words}
\title{Filter stop words from unigrams}
\usage{
filter_filler_words(tokenized_unigrams, filler_words)
}
\arguments{
\item{tokenized_unigrams}{A unigram dataframe.}

\item{filler_words}{A filler_words dataframe.}
}
\description{
Filter stop words from unigrams
}
