% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tidytext.R
\name{filter_stop_words}
\alias{filter_stop_words}
\title{Filter stop words from unigrams}
\usage{
filter_stop_words(tokenized_unigrams, stop_words)
}
\arguments{
\item{tokenized_unigrams}{A unigram dataframe.}

\item{stop_words}{A stop_words dataframe.}
}
\description{
Filter stop words from unigrams
}
