% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_twitter.R
\name{get_creds}
\alias{get_creds}
\title{Retrieve Twiter API keys}
\usage{
get_creds(
  api_key = NULL,
  api_key_secret = NULL,
  access_token = NULL,
  access_token_secret = NULL
)
}
\description{
Retrieve Twitter API keys either from .Renviron file, or users
can manually pass them into this function. If all 4 credentials are found,
this function returns a named list containing the keys. If any are missing,
this function returns a character vector containing the names of the missing
keys.
}
\details{
twitter
}
