% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tidytext.R
\name{get_lexicons}
\alias{get_lexicons}
\title{Download lexicons}
\usage{
get_lexicons(retrieve = FALSE)
}
\arguments{
\item{retrieve}{If TRUE, return a list of 3 lexicons. If FALSE, invisibly
return the list. (used for downloading the lexicons initially)}
}
\description{
Download Lexicons (AFINN, bing, nrc) to the user cache. This
function asks the user to explicitly agree the conditions for the data.
}
