% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bot.R
\name{mod_bot_ui}
\alias{mod_bot_ui}
\title{mod_bot_ui and mod_bot_name_server}
\usage{
mod_bot_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module, that contains the last step of the app, namely
the resulting analysis, and a download button. The analysis should contain
the information about the data & method feeding in, and of course, the
sentiment score.
}
\details{
bot UI Function
}
