% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_creds_modal.R
\name{mod_creds_modal_ui}
\alias{mod_creds_modal_ui}
\alias{mod_creds_modal_server}
\title{creds_modal Module}
\usage{
mod_creds_modal_ui(id)

mod_creds_modal_server(id, to_enable)
}
\arguments{
\item{id}{The Module namespace}

\item{to_enable}{a character vector of missing API credentials (input id of modal).}
}
\description{
A shiny Module containing the modalBox that asks users for
Twitter Dev API credentials.
}
