% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_header.R
\name{mod_header_ui}
\alias{mod_header_ui}
\alias{mod_header_server}
\title{Header module}
\usage{
mod_header_ui(id)

mod_header_server(id)
}
\arguments{
\item{id}{The Module namespace}
}
\description{
A shiny Module that contains the header of the app,
namely the title and the image.
}
