% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_header.R
\name{mod_header_ui}
\alias{mod_header_ui}
\title{mod_header_ui and mod_header_name_server}
\usage{
mod_header_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module, that contains the header of the app,
namely the title and the image.
}
\details{
header UI Function
}
