% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main.R
\name{mod_main_ui}
\alias{mod_main_ui}
\alias{mod_main_server}
\title{Main module}
\usage{
mod_main_ui(id)

mod_main_server(id)
}
\arguments{
\item{id}{The Module namespace}
}
\description{
This module calls everything.
}
