% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mid_lexicons.R
\name{mod_mid_lexicons_ui}
\alias{mod_mid_lexicons_ui}
\alias{mod_mid_lexicons_server}
\title{mid_lexicons module}
\usage{
mod_mid_lexicons_ui(id)

mod_mid_lexicons_server(id, ta)
}
\arguments{
\item{id}{The Module namespace}

\item{ta}{TweetAnalysis object, to hold analysis process in R6.}
}
\description{
Called from mod_mid, this module is responsible for retrieving the Lexicons
to be used for sentiment analysis. Everytime the lexicon is changed,
it directly edits the `lexicons` field in R6 right away, without trigger.
}
