% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mid_negation_words.R
\name{mod_mid_negation_words_ui}
\alias{mod_mid_negation_words_ui}
\alias{mod_mid_negation_words_server}
\title{mid_negation_words module}
\usage{
mod_mid_negation_words_ui(id)

mod_mid_negation_words_server(id, ta)
}
\arguments{
\item{id}{The Module namespace}

\item{ta}{TweetAnalysis object, to hold analysis process in R6.}
}
\description{
Called from mod_mid, this module is responsible for adding/removing/seeing
the list of negation words used for the analysis. Everytime this list is edited,
it directly edits the `negation_words` field in R6 right away, without trigger.
}
