% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_top.R
\name{mod_top_ui}
\alias{mod_top_ui}
\title{mod_top_ui and mod_top_name_server}
\usage{
mod_top_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module that contains the first step of the app,
namely the various ways to pull in tweets. Users must specifty 1 of 4 ways
to filter tweets, and hit "Pull Tweets!" button, which will pull in tweets
and display as a table. Once table is loaded, the users can move on to
the second step.
}
\details{
top UI Function
}
