% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_twitter.R
\name{pretty_tweets}
\alias{pretty_tweets}
\title{Generate DT-friendly format tweets}
\usage{
pretty_tweets(tweets)
}
\arguments{
\item{tweets}{A data.frame containing tweets.}
}
\description{
Process tweets pulled by the user, so that it looks nice on DT.
}
\details{
The resulting dataframe contains four columns
\itemize{
\item{Picture:} {User picture, with an anchor tag, leading to user profile}
\item{User:} {User name}
\item{Date:} {Date/time in UTC}
\item{Text:} {Text from tweet, wrapped in anchor tag, leading to the tweet}
}
}
