% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_twitter.R
\name{process_tweets}
\alias{process_tweets}
\title{Process tweets}
\usage{
process_tweets(tweets)
}
\arguments{
\item{tweets}{A data.frame containing tweets.}
}
\description{
Process tweets pulled by the user
}
