% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tidytext.R
\name{produce_analysis_output}
\alias{produce_analysis_output}
\title{Produce the analysis outputs}
\usage{
produce_analysis_output(dfs)
}
\arguments{
\item{dfs}{A list of tibbles, passed from \code{tokenize_tweets}. This list
contains the original tweets, as well as tokenized sentiments.}
}
\description{
Makes adjustments to assign a negative score to phrases like "I am not happy",
that would have gotten a positive score, had it not been adjusted (looking
at single word at a time).
}
