% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tryCatch_alert.R
\name{tryCatch_alert}
\alias{tryCatch_alert}
\title{tryCatch with shinyalert messages}
\usage{
tryCatch_alert(expr, input_id, type)
}
\arguments{
\item{expr}{Expression to try. In our case, add/remove words from R6 fields.}

\item{input_id}{the input ID of the text. (negation_word or filler_word)}

\item{type}{A string, that makes the shinyalert ID. (add or remove)}
}
\description{
A tryCatch with output wrapped in shinyalert.
}
