% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_R6.R
\name{TweetAnalysis}
\alias{TweetAnalysis}
\title{R6 class of tweet analyzing process}
\description{
R6 class of tweet analyzing process

R6 class of tweet analyzing process
}
\details{
R6 object to encapsulate the reading/cleaning/export of tweets
}
\section{Public fields}{

* `data`:
* `data_print`: print the data
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Pulled tweets. This is subject to change, every time a set
of tweets are pulled.}

\item{\code{adjust_negation}}{If "yes", do adjust bi-gram adjustment.}

\item{\code{negation_words}}{Negation words used for bi-gram adjustments.}

\item{\code{filler_words}}{Stop words that are irrelevant from analysis.
Dataset from tidytext.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-add_negation_word}{\code{TweetAnalysis$add_negation_word()}}
\item \href{#method-remove_negation_word}{\code{TweetAnalysis$remove_negation_word()}}
\item \href{#method-add_filler_word}{\code{TweetAnalysis$add_filler_word()}}
\item \href{#method-remove_filler_word}{\code{TweetAnalysis$remove_filler_word()}}
\item \href{#method-clone}{\code{TweetAnalysis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_negation_word"></a>}}
\if{latex}{\out{\hypertarget{method-add_negation_word}{}}}
\subsection{Method \code{add_negation_word()}}{
Add word to negation_words (lower case). Throw a warning if the word
already exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAnalysis$add_negation_word(negation_word)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{negation_word}}{Negation word}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_negation_word"></a>}}
\if{latex}{\out{\hypertarget{method-remove_negation_word}{}}}
\subsection{Method \code{remove_negation_word()}}{
Remove word from negation_words (lower case). Throw a warning if the
word doesn't exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAnalysis$remove_negation_word(negation_word)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{negation_word}}{Negation word}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_filler_word"></a>}}
\if{latex}{\out{\hypertarget{method-add_filler_word}{}}}
\subsection{Method \code{add_filler_word()}}{
Add word from filler_words (lower case). Throw a warning if the word
already exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAnalysis$add_filler_word(filler_word)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filler_word}}{Filler word}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_filler_word"></a>}}
\if{latex}{\out{\hypertarget{method-remove_filler_word}{}}}
\subsection{Method \code{remove_filler_word()}}{
Remove word from filler_words (lower case). Throw a warning if the word
doesn't exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAnalysis$remove_filler_word(filler_word)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filler_word}}{Filler word}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAnalysis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
