% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_twitter.R
\name{connect_to_api}
\alias{connect_to_api}
\title{Connect to Twitter API}
\usage{
connect_to_api(bearer_token, save = TRUE)
}
\arguments{
\item{bearer_token}{User specific bearer token}

\item{save}{Boolean to indicate whether to save the authentication or not,
on initial setup. Default to TRUE.}
}
\description{
Connect to Twitter API using the credentials provided.
}
