% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bot.R
\name{mod_bot_ui}
\alias{mod_bot_ui}
\alias{mod_bot_server}
\title{Bottom module
A shiny Module, that contains the last step of the app, namely
the resulting analysis, and a download button. The analysis output contains
the information about the data & method feeding in, and of course, the
sentiment score.}
\usage{
mod_bot_ui(id)

mod_bot_server(id, ta)
}
\arguments{
\item{id}{The Module namespace}

\item{ta}{TweetAnalysis object, to hold analysis process in R6.}
}
\description{
Bottom module
A shiny Module, that contains the last step of the app, namely
the resulting analysis, and a download button. The analysis output contains
the information about the data & method feeding in, and of course, the
sentiment score.
}
