% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_creds_modal.R
\name{mod_creds_modal_server}
\alias{mod_creds_modal_server}
\title{creds_modal Module}
\usage{
mod_creds_modal_server(id)
}
\description{
A shiny Module containing a popup box that asks users for
Twitter Dev API bearer token. Once user inputs the token, it authenticates,
saves as .rds
}
