% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mid.R
\name{mod_mid_ui}
\alias{mod_mid_ui}
\alias{mod_mid_server}
\title{Mid module}
\usage{
mod_mid_ui(id)

mod_mid_server(id, ta)
}
\arguments{
\item{id}{The Module namespace}

\item{ta}{TweetAnalysis object, to hold analysis process in R6.}
}
\description{
A shiny Module that contains the second step of the app,
namely the 2 customizable aspects of the sentiment analysis. Firstly,
Negation adjustment carries out the bi-gram adjustment to the original
uni-gram analysis. Users can provide their own negation words. Secondly,
users can specify their own stop words - words that are irrelevant to the
sentiment of a sentence ('the', 'of'). For both these options, users can
check if a word they'd like to submit already exists in the analysis by
clicking "see list" button.
}
