% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mid.R
\name{mod_mid_ui}
\alias{mod_mid_ui}
\alias{mod_mid_server}
\title{Mid module}
\usage{
mod_mid_ui(id)

mod_mid_server(id)
}
\arguments{
\item{id}{The Module namespace}
}
\description{
A shiny Module that contains the second step of the app,
namely the various methods to analyze the tweets. Users can either analyze
every word (unigram), or pairs of words (N-gram), using several different
dictionaries. If one option is selected, the other should be NULL. Once
the method has been determined, the "Analyze!" button executes the analysis.
}
