% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{TweetAnalysis}
\alias{TweetAnalysis}
\title{R6 class of tweet analyzing process}
\description{
R6 class of tweet analyzing process

R6 class of tweet analyzing process
}
\details{
R6 object to encapsulate the reading/cleaning/export of tweets
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Pulled tweets in tibble. This is subject to change, every time a set
of tweets are pulled.}

\item{\code{lexicons}}{A list of tibbles containing lexicons. It is NULL by
default, but is given the list at initialization}

\item{\code{adjust_negation}}{If "yes", do adjust bi-gram adjustment.}

\item{\code{negation_words}}{Negation words used for bi-gram adjustments.}

\item{\code{stop_words}}{A tibble of stop words that are irrelevant from
analysis. Dataset from tidytext is passed, during initialize.}

\item{\code{analysis_result}}{A list of two tibbles and a ggplot plot, that
contain the result of text analysis, performed by self$analyze()}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-add_negation_word}{\code{TweetAnalysis$add_negation_word()}}
\item \href{#method-remove_negation_word}{\code{TweetAnalysis$remove_negation_word()}}
\item \href{#method-add_stop_word}{\code{TweetAnalysis$add_stop_word()}}
\item \href{#method-remove_stop_word}{\code{TweetAnalysis$remove_stop_word()}}
\item \href{#method-analyze}{\code{TweetAnalysis$analyze()}}
\item \href{#method-print_analysis}{\code{TweetAnalysis$print_analysis()}}
\item \href{#method-new}{\code{TweetAnalysis$new()}}
\item \href{#method-clone}{\code{TweetAnalysis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_negation_word"></a>}}
\if{latex}{\out{\hypertarget{method-add_negation_word}{}}}
\subsection{Method \code{add_negation_word()}}{
Add word to negation_words (lower case). Throw a warning if the word
already exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAnalysis$add_negation_word(word)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{word}}{Negation word}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_negation_word"></a>}}
\if{latex}{\out{\hypertarget{method-remove_negation_word}{}}}
\subsection{Method \code{remove_negation_word()}}{
Remove word from negation_words (lower case). Throw a warning if the
word doesn't exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAnalysis$remove_negation_word(word)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{word}}{Negation word}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_stop_word"></a>}}
\if{latex}{\out{\hypertarget{method-add_stop_word}{}}}
\subsection{Method \code{add_stop_word()}}{
Add word from stop_words (lower case). Throw a warning if the word
already exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAnalysis$add_stop_word(word)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{word}}{Stop word}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_stop_word"></a>}}
\if{latex}{\out{\hypertarget{method-remove_stop_word}{}}}
\subsection{Method \code{remove_stop_word()}}{
Remove word from stop_words (lower case). Throw a warning if the word
doesn't exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAnalysis$remove_stop_word(word)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{word}}{Stop word}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-analyze"></a>}}
\if{latex}{\out{\hypertarget{method-analyze}{}}}
\subsection{Method \code{analyze()}}{
Conduct sentiment analysis. It calls \code{conduct_analysis} using arguments:
\itemize{
\item{self$data}
\item{self$lexicons}
\item{self$stop_words}
\item{self$negation_words}
\item{self$adjust_negation}
}
Which returns a list of tibbles. This list is fed to
\code{produce_analysis_output} to produce final outputs, and are saved into
\code{self$analysis_result}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAnalysis$analyze()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-print_analysis}{}}}
\subsection{Method \code{print_analysis()}}{
Prints out key information about the analysis, as a vector of characters
including html br tag.
\itemize{
\item{# of tweets}
\item{Overall sentiment}
\item{Average sentiment per tweet}
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAnalysis$print_analysis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Generate a new class. It requires the stop_words internal data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAnalysis$new(stop_words)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stop_words}}{stop_words internal data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TweetAnalysis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
